ClanBackend-main/requirements.txt
amqp==5.2.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
boto3==1.34.158
botocore==1.34.158
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==43.0.0
daphne==4.1.2
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.8.0rc2
deprecation==2.0.7
Django==5.1
django-autoslug==1.9.9
django-celery-beat==2.7.0
django-cities-light==3.10.1
django-cors-headers==4.4.0
django-countries==7.6.1
django-dotenv==1.4.2
django-environ==0.11.2
django-filter==24.3
django-jsonfield==1.4.1
django-minio-storage==0.5.7
django-model-utils==4.5.1
django-oauth-toolkit==2.4.0
django-parler==2.3
django-parler-rest==2.2
django-phonenumber-field==8.0.0
django-postgres-extra==2.0.8
django-select-multiple-field==0.4.2
django-storages==1.14.4
django-templated-mail==1.1.1
django-timezone-field==7.0
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-word-filter==0.2.0
docopt==0.6.2
drf-social-oauth2==3.1.0
drf-yasg==1.21.7
executing==2.0.1
fastjsonschema==2.20.0
fido2==1.1.3
fqdn==1.5.1
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
infi.clickhouse-orm==2.1.3
inflection==0.5.1
ipykernel==6.29.5
ipython==8.12.3
iso8601==2.1.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jwcrypto==1.5.6
kombu==5.4.0
Markdown==3.5.2
MarkupSafe==2.1.5
martor==1.6.44
matplotlib-inline==0.1.7
minio==7.2.8
mistune==3.0.2
msgpack==1.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
oauthlib==3.2.2
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.42
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
progressbar2==4.4.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-utils==3.8.2
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-generic-relations==2.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
Send2Trash==1.8.3
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
statsd==4.0.1
swapper==1.4.0
terminado==0.18.1
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
Twisted==24.7.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
Unidecode==1.3.8
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.11
yarg==0.1.9
zope.interface==7.0.1
